pyproject.toml
[build-system]
build-backend = "pdm.backend"
requires = [
    "pdm-backend",
]

[project]
authors = [
    {email = "louismmx@gmail.com", name = "Louis Maddox"},
    {name = "Ali Farhadi"},
    {name = "Luca Weihs"},
    {name = "Mohit Bansal"},
    {name = "Ranjay Krishna"},
    {name = "Tanmay Gupta"},
    {name = "Zaid Khan"},
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python",
]
dependencies = [
    "bm25s>=0.2.7.post1",
    "chromadb>=0.6.3",
    "click>=8.1.8",
    "colorama>=0.4.6",
    "datasets>=3.3.2",
    "gitpython>=3.1.44",
    "hydra-core>=1.3.2",
    "instructor>=1.7.2",
    "ipdb>=0.13.13",
    "jinja2>=3.1.5",
    "lancedb>=0.19.0",
    "loguru>=0.7.3",
    "networkx>=3.4.2",
    "numpy>=2.2.3",
    "omegaconf>=2.3.0",
    "openai>=1.64.0",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "qdrant-client>=1.13.2",
    "rich>=13.9.4",
    "sacrebleu>=2.5.1",
    "sqlalchemy>=2.0.38",
    "tenacity>=9.0.0",
    "tiktoken>=0.9.0",
    "together>=1.4.1",
    "tqdm>=4.67.1",
    "tree-sitter-python>=0.23.6",
    "tree-sitter>=0.24.0",
]
description = "Planning for code-use tasks using embedded intents."
keywords = [
    "code generation",
    "embedding",
    "planning",
    "repositories",
]
name = "mutagrep"
readme = "README.md"
requires-python = ">=3.10"
version = "0.1.0"

[project.license]
text = "MIT"

[project.optional-dependencies]
dev = [
    "pdm-bump>=0.9.10",
    "pdm>=2.22.3",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
]

[project.scripts]
mutagrep = "mutagrep.cli:main"

[project.urls]
Documentation = "https://mutagrep.vercel.app/"
Homepage = "https://github.com/lmmx/mutagrep"
Repository = "https://github.com/lmmx/mutagrep.git"

[tool.ruff.lint]
ignore = [
    "C408",
    "C901",
    "E501",
    "E722",
    "E741",
]
select = [
    "C",
    # "D",
    "E",
    "F",
    "I",
    "UP",
    "W",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "E401",
    "F401",
]

[tool.tomlsort]
all = true
in_place = true
spaces_before_inline_comment = 2
spaces_indent_inline_array = 4
trailing_comma_inline_array = true
